feat(vscode): add Pochi webview panel and sidebar integration with commands
