+++ packages/vscode/src/integrations/command.ts
@@ -3,7 +3,7 @@ import {
   calcEditedRangeAfterAccept,
 } from "@/code-completion/auto-code-actions";
 // biome-ignore lint/style/useImportType: needed for dependency injection
-import { RagdollWebviewProvider } from "@/integrations/webview/ragdoll-webview-provider";
+import { PochiWebviewPanel, PochiWebviewSidebar } from "@/integrations/webview";
 import type { AuthClient } from "@/lib/auth-client";
 // biome-ignore lint/style/useImportType: needed for dependency injection
 import { AuthEvents } from "@/lib/auth-events";
@@ -40,13 +40,15 @@ export class CommandManager implements vscode.Disposable {
   private disposables: vscode.Disposable[] = [];
 
   constructor(
-    private readonly ragdollWebviewProvider: RagdollWebviewProvider,
+    private readonly pochiWebviewProvider: PochiWebviewSidebar,
     private readonly newProjectRegistry: NewProjectRegistry,
     @inject("AuthClient") private readonly authClient: AuthClient,
     private readonly authEvents: AuthEvents,
     @inject("McpHub") private readonly mcpHub: McpHub,
     private readonly pochiConfiguration: PochiConfiguration,
     private readonly posthog: PostHog,
+    @inject("vscode.ExtensionContext")
+    private readonly context: vscode.ExtensionContext,
   ) {
     this.registerCommands();
   }
@@ -64,7 +66,7 @@ export class CommandManager implements vscode.Disposable {
       await prepareProject(workspaceUri, githubTemplateUrl, progress);
     }
 
-    const webviewHost = await this.ragdollWebviewProvider.retrieveWebviewHost();
+    const webviewHost = await this.pochiWebviewProvider.retrieveWebviewHost();
     webviewHost.openTask(openTaskParams);
 
     if (requestId) {
@@ -209,7 +211,7 @@ export class CommandManager implements vscode.Disposable {
             progress.report({ message: "Pochi: Opening task..." });
             await vscode.commands.executeCommand("pochiWebui.focus");
             const webviewHost =
-              await this.ragdollWebviewProvider.retrieveWebviewHost();
+              await this.pochiWebviewProvider.retrieveWebviewHost();
             webviewHost.openTask({ uid });
           },
         );
@@ -220,7 +222,7 @@ export class CommandManager implements vscode.Disposable {
         async () => {
           await vscode.commands.executeCommand("pochiWebui.focus");
           const webviewHost =
-            await this.ragdollWebviewProvider.retrieveWebviewHost();
+            await this.pochiWebviewProvider.retrieveWebviewHost();
           webviewHost.openTask({ uid: undefined });
         },
       ),
@@ -230,7 +232,7 @@ export class CommandManager implements vscode.Disposable {
         async () => {
           await vscode.commands.executeCommand("pochiWebui.focus");
           const webviewHost =
-            await this.ragdollWebviewProvider.retrieveWebviewHost();
+            await this.pochiWebviewProvider.retrieveWebviewHost();
           webviewHost.openTaskList();
         },
       ),
@@ -240,7 +242,7 @@ export class CommandManager implements vscode.Disposable {
         async () => {
           await vscode.commands.executeCommand("pochiWebui.focus");
           const webviewHost =
-            await this.ragdollWebviewProvider.retrieveWebviewHost();
+            await this.pochiWebviewProvider.retrieveWebviewHost();
           webviewHost.openSettings();
         },
       ),
@@ -307,7 +309,7 @@ export class CommandManager implements vscode.Disposable {
 
       vscode.commands.registerCommand("pochi.toggleFocus", async () => {
         const webviewHost =
-          await this.ragdollWebviewProvider.retrieveWebviewHost();
+          await this.pochiWebviewProvider.retrieveWebviewHost();
         if (await webviewHost.isFocused()) {
           logger.debug("Focused on editor");
           await vscode.commands.executeCommand(
@@ -420,6 +422,10 @@ export class CommandManager implements vscode.Disposable {
           });
         },
       ),
+
+      vscode.commands.registerCommand("pochi.openInEditor", async () => {
+        PochiWebviewPanel.createOrShow(this.context.extensionUri);
+      }),
     );
   }
