+++ packages/vscode/src/integrations/webview/index.ts
@@ -1,0 +1,2 @@
+export { PochiWebviewPanel } from "./pochi-webview-panel";
+export { PochiWebviewSidebar } from "./webview-sidebar";
